refactor(scheduler): extract maintenance batch handlers (issue #224 task 2)
